fix(frontend): apply feedbacks
