+++ packages/code-du-travail-frontend/pages/index.js
@@ -2,7 +2,7 @@ import React from "react";
 import getConfig from "next/config";
 import fetch from "isomorphic-unfetch";
 
-import SearchLanding from "../src/search/SearchLanding";
+import SearchHero from "../src/search/SearchHero";
 import { Layout } from "../src/layout/Layout";
 import Themes from "../src/common/Themes";
 import Outils from "../src/common/Outils";
@@ -13,15 +13,15 @@ const {
 } = getConfig();
 
 const Home = ({ pageUrl, ogImage, children = [] }) => (
-  <Layout hasSearch={false}>
+  <Layout hideSearch={true}>
     <Metas
       url={pageUrl}
       title="Code du travail numérique"
       description="Posez votre question sur le droit du travail et obtenez une réponse personalisée à vos questions (formation, rupture de contrat, démission, indémnités)."
       image={ogImage}
     />
-    <SearchLanding />
-    <Themes themes={children} />
+    <SearchHero />
+    <Themes isRoot={true} themes={children} />
     <Outils />
   </Layout>
 );
