+++ packages/code-du-travail-frontend/src/layout/Layout.js
@@ -3,10 +3,10 @@ import Header from "./Header.js";
 import Footer from "./Footer.js";
 import { ErrorBoundary } from "../common/ErrorBoundary";
 
-const Layout = ({ children, hasSearch = true }) => {
+const Layout = ({ children, hideSearch }) => {
   return (
     <ErrorBoundary>
-      <Header hasSearch={hasSearch} />
+      <Header hideSearch={hideSearch} />
       <main id="main">{children}</main>
       <Footer />
     </ErrorBoundary>
